+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project are documented in this file.
 ### Fixed
 - **examples** Prometheus example ([#562](https://github.com/geofmureithi/apalis/pull/562))
 
+## Fixed
+
+- **PostgresStorage**: fix type error when updating jobs ([#539](https://github.com/geofmureithi/apalis/issues/539))
+
 ## [0.7.1](https://github.com/geofmureithi/apalis/releases/tag/v0.7.1)
 
 ### Changed